[MCAnalysis] Remove the concept of "MCAtom".
This was the big problem with the MC CFG that was biting us so hard.
The concept was originally introduced to provide a separation between
MC-level control flow graphs and raw "atoms" in the binary.  One of
the advantages was that if a chunk of code is used in multiple
MCFunctions, its content isn't duplicated across several MCBasicBlocks,
but instead, they can all share the same backing MCAtom.

In practice, this isn't that useful.  The little added duplication
is tremendously offset by the fact that we don't need to maintain a
MCBasicBlock<>MCAtom mapping anymore, not a global list of Atoms.

Instead, we can just use MCBasicBlocks everywhere.