[MCAnalysis] Build MCBasicBlocks in a single step, not incrementally.
Being able to modify MCBasicBlocks wasn't used, but added quite a bit
of complexity.  Only the MCOD needs to be able to change them anyway,
so we might as well move the functionality that dealt with BB splitting
and such to MCOD, where it was partially duplicated.
Still need to change MCModuleYAML to directly build MCBBs;  this will
avoid the addInst/addPred/addSucc methods and make it truly read-only.