[MCAnalysis] Add LC_SEGMENT vmaddr to LC_MAIN file offset.
Previously, the code only intended to check for LC_MAIN when
we have a dyld-provided header address, so that we can apply
it directly to get the memory-mapped address.

A bug snuck in where LC_MAIN was completely ignored: this happened
to work in the common case because we fallback to looking for "_main".

Fix that, and teach the symbolizer to also handle the static case, by
looking for the segment that the LC_MAIN command points into.