[DC][X86] Introduce CtlSysEFLAGS to avoid unnecessary updates.
Before this, we had to assume that every EFLAGS update was live
across function boundaries, because the non-status bits had to be
inserted when updating the status bits, leading to reuse of the
whole register, with bitwise ops that are nontrivial to reason about.

Separating status from the other flags lets us remove that dependency.