[DC] Prefer largest super-registers with the least subreg units.
Multiple same-sized super-registers of a register, if they are
super-registers of multiple regunits, are usually encoding tuples and
shouldn't be taken into account when computing the largest
super-register of a given register.

For instance, on AArch64, W1 has three super-registers: X1, W0_W1, and
W1_W2.  The latter two are super-registers of multiple unit roots, thus
breaking the 'largest super-register' tree.

Instead of picking the last one, ignore those, and only pick the
super-register that participates in the least (unique) number of trees.

Also add the concept of 'aliased' register for these ignored
super-registers: they don't live in the regset, and are (will eventually
be) materialized lazily from their sub-registers.  Really, they're
'virtual' registers from the standpoint of the regset, but let's
avoid overloading terminology too much.