[DC] Declare number of results/operands in emitted Sema array.
Previously, the implementation of each operation was responsible for
manually getting the elements of the semantics array it needed, using
raw accesses to the unsigned contents.  This was very ugly, confusing,
and error-prone, as order of access to operands mattered, and errors
manifest in cryptic ways.

Instead, always emit a "signature" element, containing:
- the number of defined results in the high byte
- the number of operands in the low byte.

This lets us generically:
- extract all result MVTs from the semantics array, and convert them to
  types right then and there
- gather all operands from the Vals (previously defined values) array.

While there:
- switch to 16-bit entries in the semantics array
- add more detailed (pattern) comments in the generated .inc
- document more of DCInstruction.