[DC] Increment PC before any translation attempt.
This lets translateTargetInst() assume that the PC is already
incremented; the standard translation flow is unaffected.

In practice, this also means that instruction that are completely
unsupported (ignored by translateTargetInst, and that don't have opcodes
in the semantics table) will now be translated to a PC update only,
hence the large but repetitive test updates.

Reported by: Anthony Lineberry <anthony.lineberry@gmail.com>