[MCAnalysis] Entrypoint in ELF-Files. #9
For ELF-Files it is useless to use the entrypoint specified in
the ELF-Header. Usually it does only runtime-initialisation and
then call __libc_start_main which does not return for dagger (it
seems it jumps to an address in a register after the execution
of the function) - so all output from dagger is the same for
every ELF-File. It would be better to use the main-function if
the file was not stripped and for all other cases the user
should specify the entrypoint manually until we find a stable
way to automatically find it with dagger.

Tweaked from pull request #9 by:
Ahmed Bougacha <ahmed.bougacha@gmail.com>