[MCAnalysis] Recursively disassemble initial CFG.
Previously, when building the initial CFG, we only disassembled the
entrypoints we discovered via MCObjectSymbolizer.

Instead, disassemble recursively starting from there, like we would
do when translating in llvm-dec.  This gives us a more accurate
callgraph, and makes it identical to the llvm-dec translation callgraph.

For pure MC analysis purposes, this makes the output more usable as
well, with llvm-mccfg emitting a richer callgraph.

Fixes #18.